Reorganize tests - move math to test_math and serialization to test_serialization
